+++ src/main/java/roomescape/controller/ErrorController.java
@@ -5,8 +5,7 @@ import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.exception.EntityNotFoundException;
-import roomescape.exception.ReservationExistException;
+import roomescape.exception.*;
 
 @ControllerAdvice
 public class ErrorController {
@@ -21,6 +20,38 @@ public class ErrorController {
 
     @ExceptionHandler(ReservationExistException.class)
     public ResponseEntity<ProblemDetail> handleReservationExistException(ReservationExistException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.CONFLICT);
+        problemDetail.setDetail(e.getMessage());
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
+    }
+
+    @ExceptionHandler(PastReservationException.class)
+    public ResponseEntity<ProblemDetail> handlePastReservationException(RuntimeException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
+        problemDetail.setDetail(e.getMessage());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
+    }
+
+    @ExceptionHandler(ReservationTimeConflictException.class)
+    public ResponseEntity<ProblemDetail> handleReservationTimeConflictException(ReservationTimeConflictException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
+        problemDetail.setDetail(e.getMessage());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
+    }
+
+    @ExceptionHandler(ReservationTimeExistException.class)
+    public ResponseEntity<ProblemDetail> handleReservationTimeExistException(ReservationTimeExistException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.CONFLICT);
+        problemDetail.setDetail(e.getMessage());
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
+    }
+
+    @ExceptionHandler(ThemeExistException.class)
+    public ResponseEntity<ProblemDetail> handleReservationExistException(ThemeExistException e) {
         ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.CONFLICT);
         problemDetail.setDetail(e.getMessage());
