+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,14 +2,11 @@ package roomescape.controller;
 
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.dto.AvailableReservationTimeResponse;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
-import roomescape.exception.PastReservationException;
-import roomescape.exception.ReservationTimeConflictException;
 import roomescape.service.ReservationService;
 
 import java.time.LocalDate;
@@ -46,21 +43,5 @@ public class ReservationController {
     public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
         reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(PastReservationException.class)
-    public ResponseEntity<ProblemDetail> handlePastReservationException(RuntimeException e) {
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setDetail(e.getMessage());
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
-    }
-
-    @ExceptionHandler(ReservationTimeConflictException.class)
-    public ResponseEntity<ProblemDetail> handleReservationTimeConflictException(ReservationTimeConflictException e) {
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setDetail(e.getMessage());
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
     }
 }
