+++ src/main/java/roomescape/controller/ThemeController.java
@@ -2,12 +2,10 @@ package roomescape.controller;
 
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
-import roomescape.exception.ThemeExistException;
 import roomescape.service.ThemeService;
 
 import java.util.List;
@@ -41,13 +39,5 @@ public class ThemeController {
     public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
         themeService.deleteById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(ThemeExistException.class)
-    public ResponseEntity<ProblemDetail> handleReservationExistException(ThemeExistException e) {
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.CONFLICT);
-        problemDetail.setDetail(e.getMessage());
-
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
     }
 }
